langSmith0/pyproject.toml
[project]
name = "langsmith0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.3.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-core>=2.0.0",
]